+++ data_adapter/wishlist.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
 import uuid
 
-from pydantic_models.wishlist import WishlistCreate, WishlistUpdate
+from pydantic_models.wishlist import WishlistCreate
 from data_adapter.db_models.wishlist import Wishlist
 
 def create_new_wishlist(wishlist:WishlistCreate,db:Session):
@@ -19,17 +19,19 @@ def list_wishlists(db:Session):
 
 def delete_wishlist(wishlist_uuid:str,db:Session):
     wishlist = db.query(Wishlist).filter(Wishlist.uuid == wishlist_uuid).first()
+    if wishlist is None:
+        return False
     db.delete(wishlist)
     db.commit()
-    return wishlist
+    return True
 
 def get_wishlist(wishlist_uuid:str,db:Session):
     return db.query(Wishlist).filter(Wishlist.uuid == wishlist_uuid).first()
 
-def update_wishlist(wishlist:WishlistUpdate,db:Session):
+def update_wishlist(wishlist_uuid:str,wishlist:WishlistCreate,db:Session):
     wishlist_model = wishlist.model_dump()
-    print(wishlist_model)
-    existing_wishlist = db.query(Wishlist).filter(Wishlist.uuid == wishlist.uuid)
+    wishlist_model["uuid"] = wishlist_uuid
+    existing_wishlist = db.query(Wishlist).filter(Wishlist.uuid == wishlist_uuid)
     if existing_wishlist.first() is None:
         return False
     existing_wishlist.update(wishlist_model)
